chore(hype-ui): Cleaned up styles and components
