+++ client/package.json
@@ -70,6 +70,7 @@
     "@storybook/react": "^6.4.22",
     "@storybook/testing-library": "^0.0.11",
     "@types/react-router-dom": "^5.1.8",
+    "eslint-config-react-app": "^7.0.1",
     "react-infinite-scroll-component": "^6.1.0",
     "webpack": "^5.72.0"
   },
