--- client/src/App.css
@@ -1,43 +1,0 @@
-.App {
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-body {
-  background-color: #ececec;
-}
-
-.page-container {
-  background: #fafafa;
-  border-radius: 2rem;
-}
-
-.appContainer {
-  background: #fafafa;
-  border-radius: 2rem;
-  margin: 1.5rem 2rem 1.5rem;
-  min-height: 100vh;
-}
-
