+++ client/src/App.tsx
@@ -1,11 +1,19 @@
-import React from "react";
-import "./App.css";
-import Header from "./components/header/Header";
-import { useMediaQuery } from "react-responsive";
-import useMetamask from "./hooks/useMetamask";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import { MetaMaskProvider } from "metamask-react";
-import Home from "./pages/Home/Home";
+import React from 'react';
+import Header from './components/header/Header';
+import { useMediaQuery } from 'react-responsive';
+import useMetamask from './hooks/useMetamask';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { MetaMaskProvider } from 'metamask-react';
+import Home from './pages/Home/Home';
+import Redeem from './pages/Redeem';
+import { HypeThemeProvider } from './theme/HypeTheme';
+import styled from 'styled-components';
+
+const StyledAppContainer = styled.div`
+  border-radius: 2rem;
+  margin: 1.5rem 2rem 1.5rem;
+  min-height: 100vh;
+`;
 
 function Root() {
   const isMobile = useMediaQuery({ query: `(max-width: 760px)` });
@@ -13,34 +21,31 @@ function Root() {
 
   return (
     <div className="App">
-        <Header
-          variant={isMobile ? "mobile" : "desktop"}
-          status={
-            status === "initializing" || status === "connecting"
-              ? "notConnected"
-              : status
-          }
-          account={account}
-          onConnect={connect}
-          children={null}
-        />
-        <div className="appContainer">
-          <Router>
-            <Switch>
-              <Route exact path="/" component={Home} />
-            </Switch>
-          </Router>
-        </div>
+      <Header
+        variant={isMobile ? 'mobile' : 'desktop'}
+        status={status === 'initializing' || status === 'connecting' ? 'notConnected' : status}
+        account={account}
+        onConnect={connect}
+        children={null}
+      />
+      <StyledAppContainer>
+        <Router>
+          <Switch>
+            <Route exact path="/" component={Home} />
+            <Route exact path="/redeem" component={Redeem} />
+          </Switch>
+        </Router>
+      </StyledAppContainer>
     </div>
   );
 }
 
-function App() {
-  return (
+const App = () => (
+  <HypeThemeProvider>
     <MetaMaskProvider>
       <Root />
     </MetaMaskProvider>
-  );
-}
+  </HypeThemeProvider>
+);
 
 export default App;
