--- client/src/assets/icons/Spinner.css
@@ -1,16 +1,0 @@
-@keyframes spinner {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-.loading-spinner {
-  animation: spinner 1.5s linear infinite;
-  position: sticky;
-  left: 50%;
-  bottom: 2.5rem;
-  transform: translate(-50%, -50%);
-  margin: 0 auto;
-}
