+++ client/src/components/header/Header.stories.tsx
@@ -1,9 +1,10 @@
-import React from "react";
-import { ComponentStory } from "@storybook/react";
-import Header, { HeaderValues } from "./Header";
+import React from 'react';
+import { ComponentStory } from '@storybook/react';
+import Header from './Header';
+import { HeaderValues } from './Header.effects';
 
 export default {
-  title: "Hype/Header",
+  title: 'Hype/Header',
   component: Header,
 };
 
@@ -11,74 +12,50 @@ const Template: ComponentStory<typeof Header> = (args) => <Header {...args} />;
 
 export const MobileConnected = Template.bind({});
 MobileConnected.args = {
-  variant: "mobile",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "connected",
+  variant: 'mobile',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'connected',
   onConnect: () => {},
-  account: "0x6e3AA85dB95BBA316276a3...82aFB",
+  account: '0x6e3AA85dB95BBA316276a3...82aFB',
 };
 
 export const MobileDisconnected = Template.bind({});
 MobileDisconnected.args = {
-  variant: "mobile",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "notConnected",
+  variant: 'mobile',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'notConnected',
   onConnect: () => {},
 };
 
 export const MobileUnavailable = Template.bind({});
 MobileUnavailable.args = {
-  variant: "mobile",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "unavailable",
+  variant: 'mobile',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'unavailable',
   onConnect: () => {},
 };
 
 export const DesktopConnected = Template.bind({});
 DesktopConnected.args = {
-  variant: "desktop",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "connected",
+  variant: 'desktop',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'connected',
   onConnect: () => {},
-  account: "0x6e3AA85dB95BBA316276a3...82aFB",
+  account: '0x6e3AA85dB95BBA316276a3...82aFB',
 };
 
 export const DesktopDisconnected = Template.bind({});
 DesktopDisconnected.args = {
-  variant: "desktop",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "notConnected",
+  variant: 'desktop',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'notConnected',
   onConnect: () => {},
 };
 
 export const DesktopUnavailable = Template.bind({});
 DesktopUnavailable.args = {
-  variant: "desktop",
-  headerElements: [
-    HeaderValues.HypeFarming,
-    HeaderValues.HypePool,
-    HeaderValues.Redeem,
-  ],
-  status: "unavailable",
+  variant: 'desktop',
+  headerElements: [HeaderValues.HypeFarming, HeaderValues.HypePool, HeaderValues.Redeem],
+  status: 'unavailable',
   onConnect: () => {},
 };
