+++ client/src/components/styles/Global.tsx
@@ -1,0 +1,25 @@
+import { createGlobalStyle } from 'styled-components';
+import { HypeThemeType } from '../../theme/HypeTheme';
+
+const GlobalStyles = createGlobalStyle<{ theme: HypeThemeType }>`
+  html {
+    font-family: 'Inter';
+    font-style: normal;
+    font-size: 16px;
+    letter-spacing: -0.02em;
+  }
+
+  body {
+    margin: 0;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    background: ${({ theme }) => theme.colors.body};
+  }
+
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+  }
+`;
+
+export default GlobalStyles;
