--- client/src/index.css
@@ -1,17 +1,0 @@
-html {
-  font-family: 'Inter';
-  font-style: normal;
-  font-size: 16px;
-  letter-spacing: -0.02em;
-}
-
-body {
-  margin: 0;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
