+++ client/src/pages/Home/Home.effects.tsx
@@ -1,0 +1,115 @@
+import { useEffect, useState } from 'react';
+
+interface CardData {
+  title: string;
+  description?: string;
+  pool: number;
+  poolToken: string;
+  bonus: number;
+  bonusToken: string;
+  minReward: number;
+  rewardToken: string;
+  startDate: Date;
+  endDate: Date;
+}
+
+export const useHomeEffects = () => {
+  const [cardData, setCardData] = useState<CardData[]>([{}] as CardData[]);
+  const [titleFilter, setTitleFilter] = useState('');
+
+  useEffect(() => {
+    setCardData([
+      {
+        title: 'ApeCoin Staking Launch!',
+        description:
+          'APE is launching it’s testnet, and we’d like everyone to come & check it out! All participants will be able to claim rewards as loooooooong All participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongvvvAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongv',
+        pool: 100000,
+        poolToken: 'APE',
+        bonus: 12000,
+        bonusToken: 'TARA',
+        minReward: 1,
+        rewardToken: 'APE',
+        startDate: new Date(),
+        endDate: new Date(new Date().setDate(new Date().getDate() + 7)),
+      },
+      {
+        title: 'ApeCoin Staking Launch!',
+        description:
+          'APE is launching it’s testnet, and we’d like everyone to come & check it out! All participants will be able to claim rewards as loooooooong All participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongvvvAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongv',
+        pool: 100000,
+        poolToken: 'APE',
+        bonus: 12000,
+        bonusToken: 'TARA',
+        minReward: 1,
+        rewardToken: 'APE',
+        startDate: new Date(),
+        endDate: new Date(new Date().setDate(new Date().getDate() + 7)),
+      },
+      {
+        title: 'ApeCoin Staking Launch!',
+        description:
+          'APE is launching it’s testnet, and we’d like everyone to come & check it out! All participants will be able to claim rewards as loooooooong All participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongvvvAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongv',
+        pool: 100000,
+        poolToken: 'APE',
+        bonus: 12000,
+        bonusToken: 'TARA',
+        minReward: 1,
+        rewardToken: 'APE',
+        startDate: new Date(),
+        endDate: new Date(new Date().setDate(new Date().getDate() + 7)),
+      },
+      {
+        title: 'ApeCoin Staking Launch!',
+        description:
+          'APE is launching it’s testnet, and we’d like everyone to come & check it out! All participants will be able to claim rewards as loooooooong All participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongvvvAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongv',
+        pool: 100000,
+        poolToken: 'APE',
+        bonus: 12000,
+        bonusToken: 'TARA',
+        minReward: 1,
+        rewardToken: 'APE',
+        startDate: new Date(),
+        endDate: new Date(new Date().setDate(new Date().getDate() + 7)),
+      },
+      {
+        title: 'ApeCoin Staking Launch!',
+        description:
+          'APE is launching it’s testnet, and we’d like everyone to come & check it out! All participants will be able to claim rewards as loooooooong All participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongvvvAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongAll participants will be able to claim rewards as loooooooongv',
+        pool: 100000,
+        poolToken: 'APE',
+        bonus: 12000,
+        bonusToken: 'TARA',
+        minReward: 1,
+        rewardToken: 'APE',
+        startDate: new Date(),
+        endDate: new Date(new Date().setDate(new Date().getDate() + 7)),
+      },
+    ]);
+  }, []);
+
+  const addMoreCards = () => {
+    setTimeout(() => {
+      const currCards = cardData;
+      setCardData(currCards.concat(cardData));
+    }, 3000);
+  };
+
+  function monthDiff(dateFrom: Date, dateTo: Date): number {
+    if (!dateFrom || !dateTo) return 0;
+    return (
+      dateTo.getMonth() - dateFrom.getMonth() + 12 * (dateTo.getFullYear() - dateFrom.getFullYear())
+    );
+  }
+
+  const filteredCards = titleFilter
+    ? cardData.filter((c) => c.title.includes(titleFilter))
+    : cardData;
+
+  return {
+    setTitleFilter,
+    addMoreCards,
+    monthDiff,
+    filteredCards,
+    cardData,
+  };
+};
