+++ client/src/pages/Redeem.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Home = () => {
+  return (
+     <div>Welcome to redeem page</div>
+   );
+}
+
+export default Home;
